# backend/src/censys_ai/api/censys-ai-openapi.json
{"openapi": "3.1.0", "info": {"title": "Censys AI Summarization Agent", "description": "API for summarizing Censys cybersecurity data", "version": "0.1.0"}, "paths": {"/health": {"get": {"summary": "Health Check", "description": "Health check endpoint.", "operationId": "health_check_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/summarization/hosts": {"post": {"tags": ["summarization"], "summary": "Summarize a hosts dataset", "description": "Summarize a Censys hosts dataset.", "operationId": "summarize_hosts_summarization_hosts_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/HostsDataset"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SingleSummarizationResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/summarization/web-properties": {"post": {"tags": ["summarization"], "summary": "Summarize a web properties dataset", "description": "Summarize a Censys web properties dataset.", "operationId": "summarize_web_properties_summarization_web_properties_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/WebPropertiesDataset"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SingleSummarizationResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/summarization/certificates": {"post": {"tags": ["summarization"], "summary": "Summarize a certificates dataset", "description": "Summarize a Censys certificates dataset.", "operationId": "summarize_certificates_summarization_certificates_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/CertificatesDataset"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SingleSummarizationResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/summarization/summarize": {"post": {"tags": ["summarization"], "summary": "Summarize multiple datasets", "description": "Summarize multiple Censys datasets together.", "operationId": "summarize_summarization_summarize_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Body_summarize_summarization_summarize_post"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SummarizationResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"AutonomousSystem": {"properties": {"asn": {"type": "integer", "title": "Asn"}, "name": {"type": "string", "title": "Name"}, "country_code": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Country Code"}, "description": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Description"}, "bgp_prefix": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Bgp Prefix"}}, "type": "object", "required": ["asn", "name"], "title": "AutonomousSystem"}, "Body_summarize_summarization_summarize_post": {"properties": {"hosts": {"anyOf": [{"$ref": "#/components/schemas/HostsDataset"}, {"type": "null"}]}, "web_properties": {"anyOf": [{"$ref": "#/components/schemas/WebPropertiesDataset"}, {"type": "null"}]}, "certificates": {"anyOf": [{"$ref": "#/components/schemas/CertificatesDataset"}, {"type": "null"}]}}, "type": "object", "title": "Body_summarize_summarization_summarize_post"}, "Certificate": {"properties": {"fingerprint_sha256": {"type": "string", "title": "Fingerprint Sha256"}, "fingerprint_sha1": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Fingerprint Sha1"}, "fingerprint_md5": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Fingerprint Md5"}, "domains": {"anyOf": [{"items": {"type": "string"}, "type": "array"}, {"type": "null"}], "title": "Domains"}, "subject": {"anyOf": [{"$ref": "#/components/schemas/CertificateSubject"}, {"type": "null"}]}, "issuer": {"anyOf": [{"$ref": "#/components/schemas/CertificateIssuer"}, {"type": "null"}]}, "validity_period": {"anyOf": [{"$ref": "#/components/schemas/ValidityPeriod"}, {"type": "null"}]}, "key_info": {"anyOf": [{"$ref": "#/components/schemas/KeyInfo"}, {"type": "null"}]}, "certificate_authority": {"anyOf": [{"$ref": "#/components/schemas/CertificateAuthority"}, {"type": "null"}]}, "certificate_transparency": {"anyOf": [{"$ref": "#/components/schemas/CertificateTransparency"}, {"type": "null"}]}, "validation": {"anyOf": [{"$ref": "#/components/schemas/Validation"}, {"type": "null"}]}, "security_analysis": {"anyOf": [{"$ref": "#/components/schemas/SecurityAnalysis"}, {"type": "null"}]}, "usage_indicators": {"anyOf": [{"$ref": "#/components/schemas/UsageIndicators"}, {"type": "null"}]}}, "type": "object", "required": ["fingerprint_sha256"], "title": "Certificate"}, "CertificateAuthority": {"properties": {"name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Name"}, "type": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Type"}, "validation_level": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Validation Level"}}, "type": "object", "title": "CertificateAuthority"}, "CertificateIssuer": {"properties": {"common_name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Common Name"}, "organization": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Organization"}, "country": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Country"}}, "type": "object", "title": "CertificateIssuer"}, "CertificateSubject": {"properties": {"common_name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Common Name"}, "organization": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Organization"}, "country": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Country"}}, "type": "object", "title": "CertificateSubject"}, "CertificateTransparency": {"properties": {"logs_count": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Logs Count"}, "first_seen": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "First Seen"}, "logs": {"anyOf": [{"items": {"type": "string"}, "type": "array"}, {"type": "null"}], "title": "Logs"}}, "type": "object", "title": "CertificateTransparency"}, "CertificatesDataset": {"properties": {"metadata": {"$ref": "#/components/schemas/CertificatesDatasetMetadata"}, "certificates": {"items": {"$ref": "#/components/schemas/Certificate"}, "type": "array", "title": "Certificates"}}, "type": "object", "required": ["metadata", "certificates"], "title": "CertificatesDataset"}, "CertificatesDatasetMetadata": {"properties": {"description": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Description"}, "created_at": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Created At"}, "data_sources": {"anyOf": [{"items": {"type": "string"}, "type": "array"}, {"type": "null"}], "title": "Data Sources"}, "certificates_count": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Certificates Count"}, "certificate_fingerprints": {"anyOf": [{"items": {"type": "string"}, "type": "array"}, {"type": "null"}], "title": "Certificate Fingerprints"}}, "type": "object", "title": "CertificatesDatasetMetadata"}, "Coordinates": {"properties": {"latitude": {"type": "number", "title": "Latitude"}, "longitude": {"type": "number", "title": "Longitude"}}, "type": "object", "required": ["latitude", "longitude"], "title": "Coordinates"}, "DNS": {"properties": {"hostname": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Hostname"}}, "type": "object", "title": "DNS"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "Host": {"properties": {"ip": {"type": "string", "title": "Ip"}, "location": {"anyOf": [{"$ref": "#/components/schemas/Location"}, {"type": "null"}]}, "autonomous_system": {"anyOf": [{"$ref": "#/components/schemas/AutonomousSystem"}, {"type": "null"}]}, "whois": {"anyOf": [{"$ref": "#/components/schemas/Whois"}, {"type": "null"}]}, "dns": {"anyOf": [{"$ref": "#/components/schemas/DNS"}, {"type": "null"}]}, "services": {"anyOf": [{"items": {"$ref": "#/components/schemas/Service"}, "type": "array"}, {"type": "null"}], "title": "Services"}, "threat_intelligence": {"anyOf": [{"$ref": "#/components/schemas/ThreatIntelligence"}, {"type": "null"}]}}, "type": "object", "required": ["ip"], "title": "Host"}, "HostsDataset": {"properties": {"metadata": {"$ref": "#/components/schemas/HostsDatasetMetadata"}, "hosts": {"items": {"$ref": "#/components/schemas/Host"}, "type": "array", "title": "Hosts"}}, "type": "object", "required": ["metadata", "hosts"], "title": "HostsDataset"}, "HostsDatasetMetadata": {"properties": {"description": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Description"}, "created_at": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Created At"}, "data_sources": {"anyOf": [{"items": {"type": "string"}, "type": "array"}, {"type": "null"}], "title": "Data Sources"}, "hosts_count": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Hosts Count"}, "ips_analyzed": {"anyOf": [{"items": {"type": "string"}, "type": "array"}, {"type": "null"}], "title": "Ips Analyzed"}}, "type": "object", "title": "HostsDatasetMetadata"}, "KeyInfo": {"properties": {"algorithm": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Algorithm"}, "key_size": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Key Size"}, "public_key_fingerprint": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Public Key Fingerprint"}}, "type": "object", "title": "KeyInfo"}, "Location": {"properties": {"city": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "City"}, "country": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Country"}, "country_code": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Country Code"}, "coordinates": {"anyOf": [{"$ref": "#/components/schemas/Coordinates"}, {"type": "null"}]}, "province": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Province"}, "postal_code": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Postal Code"}, "timezone": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Timezone"}, "continent": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Continent"}}, "type": "object", "title": "Location"}, "SecurityAnalysis": {"properties": {"zlint_status": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Zlint Status"}, "failed_lints": {"anyOf": [{"items": {"type": "string"}, "type": "array"}, {"type": "null"}], "title": "Failed Lints"}, "risk_level": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Risk Level"}}, "type": "object", "title": "SecurityAnalysis"}, "Service": {"properties": {"port": {"type": "integer", "title": "Port"}, "protocol": {"type": "string", "title": "Protocol"}, "banner": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Banner"}, "banner_hash_sha256": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Banner Hash Sha256"}, "banner_hex": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Banner Hex"}, "transport_protocol": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Transport Protocol"}, "software": {"anyOf": [{"items": {"$ref": "#/components/schemas/Software"}, "type": "array"}, {"type": "null"}], "title": "Software"}, "vulnerabilities": {"anyOf": [{"items": {"$ref": "#/components/schemas/Vulnerability"}, "type": "array"}, {"type": "null"}], "title": "Vulnerabilities"}, "labels": {"anyOf": [{"items": {"type": "string"}, "type": "array"}, {"type": "null"}], "title": "Labels"}, "cert": {"anyOf": [{"additionalProperties": true, "type": "object"}, {"type": "null"}], "title": "Cert"}}, "type": "object", "required": ["port", "protocol"], "title": "Service"}, "SingleSummarizationResponse": {"properties": {"summary": {"type": "string", "title": "Summary"}}, "type": "object", "required": ["summary"], "title": "SingleSummarizationResponse"}, "Software": {"properties": {"product": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Product"}, "vendor": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Vendor"}, "version": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Version"}, "source": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Source"}, "confidence": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Confidence"}, "evidence": {"anyOf": [{"items": {"$ref": "#/components/schemas/SoftwareEvidence"}, "type": "array"}, {"type": "null"}], "title": "Evidence"}, "type": {"anyOf": [{"items": {"type": "string"}, "type": "array"}, {"type": "null"}], "title": "Type"}, "part": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Part"}, "cpe": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Cpe"}, "components": {"anyOf": [{"items": {"type": "string"}, "type": "array"}, {"type": "null"}], "title": "Components"}}, "type": "object", "title": "Software"}, "SoftwareEvidence": {"properties": {"data_path": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Data Path"}, "found_value": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Found Value"}, "literal_match": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Literal Match"}}, "type": "object", "title": "SoftwareEvidence"}, "SummarizationResponse": {"properties": {"hosts_summary": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Hosts Summary"}, "web_properties_summary": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Web Properties Summary"}, "certificates_summary": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Certificates Summary"}, "summary": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Summary"}}, "type": "object", "title": "SummarizationResponse"}, "ThreatIntelligence": {"properties": {"security_labels": {"anyOf": [{"items": {"type": "string"}, "type": "array"}, {"type": "null"}], "title": "Security Labels"}, "risk_level": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Risk Level"}}, "type": "object", "title": "ThreatIntelligence"}, "UsageIndicators": {"properties": {"ever_seen_in_scan": {"anyOf": [{"type": "boolean"}, {"type": "null"}], "title": "Ever Seen In Scan"}}, "type": "object", "title": "UsageIndicators"}, "Validation": {"properties": {"trusted_by_major_browsers": {"anyOf": [{"type": "boolean"}, {"type": "null"}], "title": "Trusted By Major Browsers"}, "validation_paths": {"anyOf": [{"$ref": "#/components/schemas/ValidationPaths"}, {"type": "null"}]}}, "type": "object", "title": "Validation"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}, "ValidationPaths": {"properties": {"apple": {"anyOf": [{"type": "boolean"}, {"type": "null"}], "title": "Apple"}, "chrome": {"anyOf": [{"type": "boolean"}, {"type": "null"}], "title": "Chrome"}, "microsoft": {"anyOf": [{"type": "boolean"}, {"type": "null"}], "title": "Microsoft"}, "mozilla": {"anyOf": [{"type": "boolean"}, {"type": "null"}], "title": "Mozilla"}}, "type": "object", "title": "ValidationPaths"}, "ValidityPeriod": {"properties": {"not_before": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Not Before"}, "not_after": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Not After"}, "length_days": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Length Days"}, "status": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Status"}}, "type": "object", "title": "ValidityPeriod"}, "Vulnerability": {"properties": {"cve_id": {"type": "string", "title": "Cve Id"}, "severity": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Severity"}, "cvss_score": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Cvss Score"}, "description": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Description"}}, "type": "object", "required": ["cve_id"], "title": "Vulnerability"}, "WebEndpoint": {"properties": {"endpoint_type": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Endpoint Type"}, "path": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Path"}, "banner": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Banner"}, "banner_hash_sha256": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Banner Hash Sha256"}, "banner_hex": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Banner Hex"}, "transport_protocol": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Transport Protocol"}}, "additionalProperties": true, "type": "object", "title": "WebEndpoint"}, "WebPropertiesDataset": {"properties": {"metadata": {"$ref": "#/components/schemas/WebPropertiesDatasetMetadata"}, "web_properties": {"items": {"$ref": "#/components/schemas/WebProperty"}, "type": "array", "title": "Web Properties"}}, "type": "object", "title": "WebPropertiesDataset"}, "WebPropertiesDatasetMetadata": {"properties": {"description": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Description"}, "created_at": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Created At"}, "data_sources": {"anyOf": [{"items": {"type": "string"}, "type": "array"}, {"type": "null"}], "title": "Data Sources"}, "web_properties_count": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Web Properties Count"}, "hostnames_analyzed": {"anyOf": [{"items": {"type": "string"}, "type": "array"}, {"type": "null"}], "title": "Hostnames Analyzed"}}, "type": "object", "title": "WebPropertiesDatasetMetadata"}, "WebProperty": {"properties": {"port": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Port"}, "scan_timed": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Scan Timed"}, "hostname": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Hostname"}, "labels": {"anyOf": [{"items": {"type": "string"}, "type": "array"}, {"type": "null"}], "title": "Labels"}, "endpoints": {"anyOf": [{"items": {"$ref": "#/components/schemas/WebEndpoint"}, "type": "array"}, {"type": "null"}], "title": "Endpoints"}, "misconfigs": {"anyOf": [{"items": {"additionalProperties": true, "type": "object"}, "type": "array"}, {"type": "null"}], "title": "Misconfigs"}, "exposures": {"anyOf": [{"items": {"additionalProperties": true, "type": "object"}, "type": "array"}, {"type": "null"}], "title": "Exposures"}, "vulns": {"anyOf": [{"items": {"additionalProperties": true, "type": "object"}, "type": "array"}, {"type": "null"}], "title": "Vulns"}, "software": {"anyOf": [{"items": {"additionalProperties": true, "type": "object"}, "type": "array"}, {"type": "null"}], "title": "Software"}, "hardware": {"anyOf": [{"items": {"additionalProperties": true, "type": "object"}, "type": "array"}, {"type": "null"}], "title": "Hardware"}, "operating_system": {"anyOf": [{"items": {"additionalProperties": true, "type": "object"}, "type": "array"}, {"type": "null"}], "title": "Operating System"}, "tls": {"anyOf": [{"items": {"additionalProperties": true, "type": "object"}, "type": "array"}, {"type": "null"}], "title": "Tls"}}, "type": "object", "title": "WebProperty"}, "Whois": {"properties": {"network": {"anyOf": [{"$ref": "#/components/schemas/WhoisNetwork"}, {"type": "null"}]}}, "type": "object", "title": "Whois"}, "WhoisNetwork": {"properties": {"handle": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Handle"}, "name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Name"}, "cidrs": {"anyOf": [{"items": {"type": "string"}, "type": "array"}, {"type": "null"}], "title": "Cidrs"}, "updated": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Updated"}}, "type": "object", "title": "WhoisNetwork"}}}}